Windows: Set $WINDOWID in Neovim process
From issue #64 - in Neovim the v:windowid var is not available, while in Unix
systems a workaround the issue is to use the $WINDOWID variable, in Windows
there is no alternative. This commit sets the $WINDOWID environment var inside
the attached Neovim process using a call to vim_command(let $WINDOWID="...").