GUI: Handle Gui Foreground notifications
For now the foreground() function in Neovim is a NOOP, implement
a replacement using notifications - use the following call to raise
the window to the front.

    call rpcnotify(0, 'Gui', 'Foreground')

The should be raised into the foreground and focused, but this is not
guaranteed to work in all systems, since the window manager may prevent
it.